Move Filesystem.scala to geotrellis-util
