+++ engine/src/main/scala/geotrellis/engine/ArgFileRasterLayer.scala
@@ -17,8 +17,8 @@
 package geotrellis.engine
 
 import geotrellis.raster._
-import geotrellis.raster.io.Filesystem
 import geotrellis.raster.io.arg.ArgReader
+import geotrellis.util.Filesystem
 
 import com.typesafe.config.Config
