+++ raster/src/main/scala/geotrellis/raster/io/geotiff/reader/TiffTagsReader.scala
@@ -1,16 +1,17 @@
 package geotrellis.raster.io.geotiff.reader
 
-import geotrellis.raster.io.Filesystem
 import geotrellis.raster.io.geotiff.tags._
 import geotrellis.raster.io.geotiff.tags.codes._
 import TagCodes._
 import TiffFieldType._
+import geotrellis.util.Filesystem
 
 import geotrellis.raster.io.geotiff.utils._
 import spire.syntax.cfor._
 import monocle.syntax._
 
 import java.nio.{ ByteBuffer, ByteOrder }
+
 
 object TiffTagsReader {
   def read(path: String): TiffTags = read(Filesystem.slurp(path))
