+++ spark/src/main/scala/geotrellis/spark/io/file/FileLayerDeleter.scala
@@ -5,7 +5,7 @@ import geotrellis.spark.io._
 import geotrellis.spark.io.avro._
 import geotrellis.spark.io.index._
 import geotrellis.spark.io.json._
-import geotrellis.raster.io.Filesystem
+import geotrellis.util.Filesystem
 import AttributeStore.Fields
 
 import spray.json.JsonFormat
@@ -13,6 +13,7 @@ import org.apache.avro.Schema
 
 import scala.reflect.ClassTag
 import java.io.File
+
 
 object FileLayerDeleter {
   def apply[K: JsonFormat: ClassTag, V: ClassTag, M: JsonFormat](attributeStore: FileAttributeStore): LayerDeleter[LayerId] =
