+++ spark/src/main/scala/geotrellis/spark/io/file/FileLayerReindexer.scala
@@ -5,7 +5,7 @@ import geotrellis.spark.io._
 import geotrellis.spark.io.avro._
 import geotrellis.spark.io.index._
 import geotrellis.spark.io.json._
-import geotrellis.raster.io.Filesystem
+import geotrellis.util.Filesystem
 import AttributeStore.Fields
 
 import org.apache.spark.SparkContext
@@ -14,6 +14,7 @@ import org.apache.avro.Schema
 
 import scala.reflect.ClassTag
 import java.io.File
+
 
 object FileLayerReindexer {
   def apply[
