+++ spark/src/main/scala/geotrellis/spark/io/file/FileRDDReader.scala
@@ -6,7 +6,7 @@ import geotrellis.spark.io.index.{MergeQueue, KeyIndex, IndexRanges}
 import geotrellis.spark.io.avro.{AvroEncoder, AvroRecordCodec}
 import geotrellis.spark.utils.KryoWrapper
 import geotrellis.spark.utils.cache.Cache
-import geotrellis.raster.io.Filesystem
+import geotrellis.util.Filesystem
 
 import org.apache.avro.Schema
 import org.apache.commons.io.IOUtils
@@ -17,6 +17,7 @@ import spire.syntax.cfor._
 import scala.collection.mutable
 import scala.reflect.ClassTag
 import java.io.File
+
 
 class FileRDDReader[K: Boundable: AvroRecordCodec: ClassTag, V: AvroRecordCodec: ClassTag](implicit sc: SparkContext) {
