+++ spark/src/main/scala/geotrellis/spark/io/file/FileRDDWriter.scala
@@ -1,13 +1,14 @@
 package geotrellis.spark.io.file
 
-import geotrellis.raster.io.Filesystem
 import geotrellis.spark.io.avro.{AvroRecordCodec, AvroEncoder}
 import geotrellis.spark.io.avro.codecs.KeyValueRecordCodec
+import geotrellis.util.Filesystem
 
 import org.apache.spark.rdd.RDD
 
 import scala.reflect.ClassTag
 import java.io.File
+
 
 class FileRDDWriter [K: AvroRecordCodec: ClassTag, V: AvroRecordCodec: ClassTag]() {
