+++ spark/src/main/scala/geotrellis/spark/io/slippy/SlippyTileReader.scala
@@ -2,12 +2,12 @@ package geotrellis.spark.io.slippy
 
 import geotrellis.vector._
 import geotrellis.raster._
-import geotrellis.raster.io.Filesystem
 import geotrellis.raster.io.geotiff._
 import geotrellis.spark._
 import geotrellis.spark.io.s3._
 import geotrellis.spark.io.hadoop._
 import geotrellis.spark.io.hadoop.formats._
+import geotrellis.util.Filesystem
 
 import org.apache.commons.io.FileUtils
 import org.apache.commons.io.filefilter._
@@ -20,6 +20,7 @@ import org.apache.hadoop.fs.Path
 
 import java.io.File
 import scala.collection.JavaConversions._
+
 
 trait SlippyTileReader[T] {
   def read(zoom: Int)(implicit sc: SparkContext): RDD[(SpatialKey, T)]
