+++ util/src/main/scala/geotrellis/util/Filesystem.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package geotrellis.raster.io
+package geotrellis.util
 
 import java.nio.file._
 import java.nio.charset.StandardCharsets
